Added test-karma.
