+++ gulpfile.js
@@ -40,7 +40,6 @@ Gulp.registry(new GulpForwardReference());
     'dashboards/scripts-dts',
     'dashboards/scripts-watch',
     'dashboards/test',
-    'dashboards/test-karma',
     'default',
     'dist',
     'dist-clean',
@@ -87,8 +86,9 @@ Gulp.registry(new GulpForwardReference());
     'scripts-watch',
     'scripts-webpacks',
     'test',
-    'test-tree',
     'test-docs',
+    'test-karma',
+    'test-tree',
     'test-ts',
     'reset-visual-references',
     'tsdoc',
