+++ tools/gulptasks/dashboards/test.js
@@ -15,10 +15,11 @@ const gulp = require('gulp');
 
 const { lintDTS } = require('../lint-dts');
 const { lintTS } = require('../lint-ts');
+const { testKarma } = require('../test-karma');
 
 gulp.task('dashboards/test', gulp.series(
     'dashboards/scripts',
     () => lintTS({ dashboards: true }),
-    'dashboards/test-karma',
+    () => testKarma({ dashboards: true }),
     () => lintDTS({ dashboards: true })
 ));
